Add input property for viewHeight.
This is to allow use of a given/computed viewHeight property rather than calculating internally. This is to resolve an issue where the parentScroll is the window, and the body content is larger than the window, but you want to use the window size as the viewHeight rather than body.clientHeight.

E.g. body.clientHeight may be 3000px and the window.innerHeight may only be 1000px. You really want to use the 1000px window height rather than the 3000px body height in the calculations for determining how many items to render.